feat(api): Move SharedProfileForCampaign to read-models folder
