+++ api/tests/unit/infrastructure/serializers/jsonapi/shared-profile-for-campaign-serializer_test.js
@@ -1,5 +1,5 @@
 const { expect, domainBuilder, sinon } = require('../../../../test-helper');
-const SharedProfileForCampaign = require('../../../../../lib/domain/models/SharedProfileForCampaign');
+const SharedProfileForCampaign = require('../../../../../lib/domain/read-models/SharedProfileForCampaign');
 const serializer = require('../../../../../lib/infrastructure/serializers/jsonapi/shared-profile-for-campaign-serializer');
 
 describe('Unit | Serializer | JSONAPI | shared-profile-for-campaign-serializer', function () {
